+++ sdk/merchantInstallment/merchantInstallment.js
@@ -3,15 +3,13 @@ const check = require('starkcore').check;
 const Resource = require('starkcore').Resource;
 
 class MerchantInstallment extends Resource {
-
     /**
      * Check out our API Documentation at https://starkbank.com/docs/api#merchant-installment
      */
 
     constructor({
-                amount, fee, fundingType, network, purchaseId, status,
-                transactionIds, tags,  due, created, updated
-                }) {
+        amount, fee, fundingType, network, purchaseId, status, transactionIds, tags,  due, created, updated
+    }) {
         super();
         this.amount = amount;
         this.fee = fee;
@@ -28,13 +26,13 @@ class MerchantInstallment extends Resource {
 }
 
 exports.MerchantInstallment = MerchantInstallment;
-let resource = {'class': exports.MerchantInstallment, 'name': 'merchantInstallment'};
+let resource = {'class': exports.MerchantInstallment, 'name': 'MerchantInstallment'};
 
 exports.get = async function (id, {user} = {}) {
     return rest.getId(resource, id, user);
 }
 
-exports.query = async function ({ limit, after, before, status, user} = {}) {
+exports.query = async function ({limit, after, before, status, user} = {}) {
     let query = {
         limit: limit,
         after: check.date(after),
@@ -44,7 +42,7 @@ exports.query = async function ({ limit, after, before, status, user} = {}) {
     return rest.getList(resource, query, user);
 }
 
-exports.page = async function ({ cursor, limit, after, before, status, user} = {}) {
+exports.page = async function ({cursor, limit, after, before, status, user} = {}) {
     let query = {
         cursor: cursor,
         limit: limit,